requirements.txt
ipykernel
albumentationsx
python-dotenv
wandb
opencv-python
torch 
torchvision
PyYAML
tqdm
pillow
transformers